hello_world/pyproject.toml
[project]
name = "hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.56.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
]
